backend/utils/parsers.py
from typing import Dict, Any
from bs4 import BeautifulSoup
import json


class DocumentParser:
    @staticmethod
    def parse_html(html: str) -> Dict[str, Any]:
        """
        Parse HTML content and extract relevant information
        """
        soup = BeautifulSoup(html, "html.parser")
        # Implement HTML parsing logic
        return {}

    @staticmethod
    def parse_pdf(pdf_content: bytes) -> Dict[str, Any]:
        """
        Parse PDF content and extract relevant information
        """
        # Implement PDF parsing logic
        return {}

    @staticmethod
    def parse_json(json_str: str) -> Dict[str, Any]:
        """
        Parse JSON content safely
        """
        try:
            return json.loads(json_str)
        except json.JSONDecodeError:
            return {}
