backend/requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Database and Models
pydantic>=2.0.0

# HTTP and Scraping
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
validators>=0.20.0

# Caching
cachetools>=5.3.0
aiohttp-client-cache>=0.8.1
aiosqlite>=0.20.0

# LLM Integration
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
